Added a (hopefully) more robust ground check
As the player contacts other objects, those objects are added to a list. The ground check looks at each item in the list, and then checks that the collision in question is below the player. So long as one of these collisions is directly below the player, the player is allowed to jump.