Made ground check much stricter
Due to the way I changed the ground check, the flag that determines whether or not I can jump was only ever cleared when I actually try to jump. As such, if I were to, say, fall off of a platform, I would be able to jump in mid-air. Now, while certain games do let you do this, I'm not too sure it would be the best design decision for this game. As such, I have added an additional check that, at the end of each frame, clears the grounded flag if you move vertically. Of course, if we were to add moving platforms, this could raise an issue, but we'll cross that bridge when we come to it.