Started reassessing input controls
While attempting to fix the issues caused by the controller lock upon death, I found out that we don't need to rely on an onMovement() function, as we can simply check the value of the "Movement" input as a float. This also means that we can remove the "Press and Release" interactions from the keyboard movement keys, as that can cause an additional lock-up upon respawning.