Removed unused scripts
Since we're not using CPU players, the Bot script is unused, and it keeps showing up in the log due to having an unused variable.
CameraFollow is unnecessary due to my CameraController script, which allows for multiple cameras to stay within the bounds of the map, even as more players join.
The Knockback script involves too much overhead with regards to timing the hitboxes when compared to my own scripts that are locked to the animation.
Finally, CharacterController is obsoleted by my own Player.cs script.